fix: remove requestedDate property
This was unneeded, and led to errors. The model of a new requested item by default has a requestedDate property which by default has a value of Date.now(), which is what we need. I believe I was trying for some reason to save the local users time when they made new items. However since people live in different time zones that play in this group self found league I need to center saving dates as UTC, and convert them to that persons local time when they view them.