feat: add requestedDate property
This property has existed on models of requested items before now however it was set by default in the model, but that was leading to undesired behavior. By moving the Date constructor to this controller function we have reached desired behavior.

fix: change date.now to new Date in markComplete

This method works better for what we want moving forward.