feat: omit template strings for dates
The previous version of these files attempted to use methods to convert stored date objects to local date and times to the users time zone, however upon further testing I realized that the views were being rendered on the server and then sent to the users. Now we just render the date objects and convert them via client side scripting.